Use curl to download instead of cohttp
This moves the HTTP and TLS concerns out of this plugin. Using curly
(which shells out to curl) is reasonably portable and is what
dune-release does.

Down the line we might want to even use `OpamDownload` since it's
guaranteed to be available, but the changes in that PR are compatible
with that approach too.

This also includes a small client for locally testing the github client.

Closes #9